requirements.txt
anyio==4.2.0
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-dotenv==1.0.0
python-telegram-bot==20.7
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
typing_extensions==4.9.0
urllib3==2.1.0
